Resolve "Store information needs to be sent to server to help customer AI select a store"

Closes #11

See merge request cs309/fall2019/tc_3_crimedime-tycoon!26

+++ Backend/src/GameClasses/Item.java
@@ -1,23 +1,21 @@
 package GameClasses;
 
-import java.math.BigDecimal;
-
 public class Item implements Comparable<Item> {
     private String name;
     private int quantity;
-    private BigDecimal wholesaleCost;
-    private BigDecimal retailCost;
+    private Double wholesaleCost;
+    private Double retailCost;
 
     public Item() {
         name = "";
-        wholesaleCost = BigDecimal.valueOf(0);
-        retailCost = BigDecimal.valueOf(0);
+        wholesaleCost = 0.0;
+        retailCost = 0.0;
     }
 
     public Item(String initName) {
         name = initName;
-        wholesaleCost = BigDecimal.valueOf(0);
-        retailCost = BigDecimal.valueOf(0);
+        wholesaleCost = 0.0;
+        retailCost = 0.0;
     }
 
     public Item(Item originalItem) {
@@ -57,7 +55,7 @@ public class Item implements Comparable<Item> {
         quantity -= subtractedItems;
     }
 
-    public BigDecimal getWholesaleCost() {
+    public Double getWholesaleCost() {
         return wholesaleCost;
     }
 
@@ -65,10 +63,10 @@ public class Item implements Comparable<Item> {
         if(newWholesaleCost < 0) {
             return;
         }
-        wholesaleCost = BigDecimal.valueOf(newWholesaleCost);
+        wholesaleCost = newWholesaleCost;
     }
 
-    public BigDecimal getRetailCost() {
+    public Double getRetailCost() {
         return retailCost;
     }
 
@@ -76,7 +74,7 @@ public class Item implements Comparable<Item> {
         if(newRetailCost < 0) {
             return;
         }
-        retailCost = BigDecimal.valueOf(newRetailCost);
+        retailCost = newRetailCost;
     }
 
     public void clearQuantity() {
